parent 2004a7b8ff9a7c153e6acb2bd13bacae00b32a3a author Matthew Carey <matthew.carey@htmcts.net> 1687189770 +0100 committer Matthew Carey <matthew.carey@htmcts.net> 1689258508 +0100
parent 2004a7b8ff9a7c153e6acb2bd13bacae00b32a3a
author Matthew Carey <matthew.carey@htmcts.net> 1687189770 +0100
committer Matthew Carey <matthew.carey@htmcts.net> 1689258278 +0100

parent 2004a7b8ff9a7c153e6acb2bd13bacae00b32a3a
author Matthew Carey <matthew.carey@htmcts.net> 1687189770 +0100
committer Matthew Carey <matthew.carey@htmcts.net> 1689253470 +0100

parent 2004a7b8ff9a7c153e6acb2bd13bacae00b32a3a
author Matthew Carey <matthew.carey@htmcts.net> 1687189770 +0100
committer Matthew Carey <matthew.carey@htmcts.net> 1689249546 +0100

initial commit based on earlier PR and Filippo's PR

only one bean this time

remove wrong whitespace

try again to fix

Still wrong but perhaps on the way

even worse but illustrates my dilemma

RIA-5804 Fixing some dependencies in DocumentManagementUploader

- Fixed dependency injection
- Fixed Sonar by adding tests

whitespace change

Fix functional tests (#588)

* Fix functional tests

* Fix functional tests

changes for CASE_DOCUMENT_AM_URL

fix values.yaml

start dealing with downloads

More download fixes

test feature flag on download client

update systemDocumentManagementUpload util files

try to fix integration test issue

fix pmd issue with integration test

comment the deprecation

appease the checkstyle gods

RIA-7385 Removed pedantic checkstyle rules (#589)

appease the checkstyle gods, more burnt offerings

appease the checkstyle gods, yet more burnt offerings

mid debugging issues with token generation in functional test

checkstyle fixes

RIA-5804 Fix attempt #1

I'm also joining in appeasing the checkstyle gods...

More appeasing the checkstyle gods, plus some minor changes

More appeasing the checkstyle gods

Attempt at config cleanup

Checkstyle...

Attempt at fixing unwise config choices

Checkstyle, of course...

mad experiment

further investigation

checkstyle issue

checkstyle issue 2

remove pointless printLn

RIA-5804 Removed commented code in a test

remove wrong whitespace

try again to fix

Still wrong but perhaps on the way

even worse but illustrates my dilemma

RIA-5804 Fixing some dependencies in DocumentManagementUploader

- Fixed dependency injection
- Fixed Sonar by adding tests

whitespace change

changes for CASE_DOCUMENT_AM_URL

fix values.yaml

start dealing with downloads

More download fixes

test feature flag on download client

update systemDocumentManagementUpload util files

try to fix integration test issue

fix pmd issue with integration test

comment the deprecation

RIA-7385 Removed pedantic checkstyle rules (#589)

appease the checkstyle gods, more burnt offerings

appease the checkstyle gods, yet more burnt offerings

mid debugging issues with token generation in functional test

checkstyle fixes

I'm also joining in appeasing the checkstyle gods...

More appeasing the checkstyle gods, plus some minor changes

More appeasing the checkstyle gods

Attempt at config cleanup

Checkstyle...

Checkstyle, of course...

mad experiment

further investigation

checkstyle issue 2

remove pointless printLn

RIA-5804 Removed commented code in a test

RIA-7384 Enforce conventions on commit messages and branches (#591)
