+++ Jenkinsfile_CNP
@@ -59,6 +59,7 @@ withPipeline(type, product, component) {
     env.DM_URL = "http://dm-store-aat.service.core-compute-aat.internal"
     env.CCD_GW_URL = "http://dm-store-aat.service.core-compute-aat.internal"
     env.IDAM_URL = "https://idam-api.aat.platform.hmcts.net"
+    env.CASE_DOCUMENT_AM_URL = "http://ccd-case-document-am-api-aat.service.core-compute-aat.internal"
     env.S2S_URL = "http://rpe-service-auth-provider-aat.service.core-compute-aat.internal"
     env.EM_BUNDLER_URL = "http://em-ccd-orchestrator-aat.service.core-compute-aat.internal"
     env.EM_BUNDLER_STITCH_URI = "/api/stitch-ccd-bundles"
@@ -74,6 +75,11 @@ withPipeline(type, product, component) {
 
     afterSuccess('functionalTest:aat') {
         steps.archiveArtifacts allowEmptyArchive: true, artifacts: '**/site/serenity/**/*'
+    }
+
+    afterAlways('test') {
+        steps.archiveArtifacts allowEmptyArchive: true, artifacts: '**/build/reports/tests/**'
+        steps.archiveArtifacts allowEmptyArchive: true, artifacts: '**/build/reports/pmd/**'
     }
 
     enableSlackNotifications('#ia-tech')
