+++ config/owasp/suppressions.xml
@@ -20,12 +20,22 @@
         ]]</notes>
         <cve>CVE-2022-45688</cve>
     </suppress>
-    <suppress>
+    <suppress until="2023-07-29">
         <cve>CVE-2023-20862</cve>
         <cve>CVE-2023-20860</cve>
         <cve>CVE-2023-20863</cve>
         <cve>CVE-2023-20873</cve>
         <cve>CVE-2023-28709</cve>
+        <cve>CVE-2023-20883</cve>
+        <cve>CVE-2023-35116</cve>
+        <cve>CVE-2023-2976</cve>
+    </suppress>
+    <suppress until="2024-06-01">
+        <cve>CVE-2023-2976</cve>
+        <cve>CVE-2020-5408</cve>
+        <cve>CVE-2023-1370</cve>
+        <cve>CVE-2023-35116</cve>
+        <cve>CVE-2020-8908</cve>
     </suppress>
     <suppress until="2024-06-01">
         <cve>CVE-2023-2976</cve>
