+++ src/functionalTest/resources/templates/minimal-aip-appeal-submitted.json
@@ -24,7 +24,7 @@
   "hasOtherAppeals": "No",
   "appellantNameForDisplay": "Talha Awan",
   "appealReferenceNumber": "PA/12345/2018",
-  "legalRepresentativeName": "A Legal Rep",
+  "legalRepresentativeName": "A Test Legal Rep",
   "legalRepresentativeEmailAddress": "{$TEST_LAW_FIRM_A_USERNAME}",
   "sendDirectionActionAvailable": "Yes",
   "hearingCentre": "taylorHouse",
