+++ src/functionalTest/resources/templates/minimal-appeal-submitted.json
@@ -26,7 +26,7 @@
   "hasOtherAppeals": "No",
   "appellantNameForDisplay": "Talha Awan",
   "appealReferenceNumber": "PA/11111/2019",
-  "legalRepresentativeName": "A Legal Rep",
+  "legalRepresentativeName": "A Test Legal Rep",
   "legalRepresentativeEmailAddress": "{$TEST_LAW_FIRM_A_USERNAME}",
   "sendDirectionActionAvailable": "Yes"
 }
