+++ src/main/java/uk/gov/hmcts/reform/iacasedocumentsapi/infrastructure/clients/DocumentManagementUploader.java
@@ -10,6 +10,7 @@ import uk.gov.hmcts.reform.iacasedocumentsapi.domain.service.FeatureToggler;
 @Service
 @RequiredArgsConstructor
 public class DocumentManagementUploader implements DocumentUploader {
+
     private final FeatureToggler featureToggler;
     private final DmDocumentManagementUploader dmDocumentManagementUploader;
     private final CdamDocumentManagementUploader cdamDocumentManagementUploader;
