--- src/main/java/uk/gov/hmcts/reform/iacasedocumentsapi/infrastructure/config/UserDetailsProviderConfiguration.java
@@ -1,25 +1,0 @@
-package uk.gov.hmcts.reform.iacasedocumentsapi.infrastructure.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.UserDetailsProvider;
-import uk.gov.hmcts.reform.iacasedocumentsapi.infrastructure.clients.IdamApi;
-import uk.gov.hmcts.reform.iacasedocumentsapi.infrastructure.security.RequestUserAccessTokenProvider;
-import uk.gov.hmcts.reform.iacasedocumentsapi.infrastructure.security.idam.IdamUserDetailsProvider;
-
-@Configuration
-public class UserDetailsProviderConfiguration {
-
-    @Bean("requestUser")
-    @Primary
-    public UserDetailsProvider getRequestUserDetailsProvider(
-        RequestUserAccessTokenProvider requestUserAccessTokenProvider,
-        IdamApi idamApi
-    ) {
-        return new IdamUserDetailsProvider(
-            requestUserAccessTokenProvider,
-            idamApi
-        );
-    }
-}
