+++ src/main/java/uk/gov/hmcts/reform/iacasedocumentsapi/infrastructure/security/RequestUserAccessTokenProvider.java
@@ -1,13 +1,11 @@
 package uk.gov.hmcts.reform.iacasedocumentsapi.infrastructure.security;
 
 import java.util.Optional;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
 
 @Service
-@Qualifier("requestUser")
 public class RequestUserAccessTokenProvider implements AccessTokenProvider {
 
     private static final String AUTHORIZATION = "Authorization";
